feat: Shape config is now loaded from a config.ini file

+++ tasks/02.images_and_colors/shape/include/shape/geometry/config/BatmanShapeConfig.h
@@ -13,7 +13,7 @@
 struct BatmanShapeConfig {
   Point2f origin;
   Point2f ovalRadius;
-  double scale{};
+  float scale{};
 };
 
 #endif /* SHAPE_BATMANSHAPECONFIG_H_ */
